+++ capi/example/examples/gain_stm.h
@@ -20,12 +20,12 @@
 void* gain_stm(void* autd) {
   void* s = NULL;
   AUTDCreateSilencer(&s, 0xFFFF, 4096);
-  AUTDSend(autd, s, NULL);
+  AUTDSend(autd, s, NULL, 20ULL * 1000ULL * 1000ULL);
   AUTDDeleteSilencer(s);
 
   void* m = NULL;
   AUTDModulationStatic(&m, 0xFF);
-  AUTDSend(autd, m, NULL);
+  AUTDSend(autd, m, NULL, 20ULL * 1000ULL * 1000ULL);
 
   double x = 90.0;
   double y = 70.0;
@@ -51,7 +51,7 @@ void* gain_stm(void* autd) {
   const double actual_freq = AUTDSTMSetFrequency(stm, 1.0);
   printf("Actual frequency is %lf Hz\n", actual_freq);
 
-  AUTDSend(autd, NULL, stm);
+  AUTDSend(autd, NULL, stm, 20ULL * 1000ULL * 1000ULL);
 
   AUTDDeleteSTM(stm);
   AUTDDeleteModulation(m);