+++ examples/soem.cpp
@@ -3,7 +3,7 @@
 // Created Date: 16/05/2022
 // Author: Shun Suzuki
 // -----
-// Last Modified: 31/01/2023
+// Last Modified: 07/03/2023
 // Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
 // -----
 // Copyright (c) 2022 Shun Suzuki. All rights reserved.
@@ -38,8 +38,6 @@ int main() try {
                   .build();
 
   auto autd = autd3::Controller::open(std::move(geometry), std::move(link));
-
-  autd.set_ack_check_timeout(std::chrono::milliseconds(20));
 
   return run(autd);
 } catch (std::exception& e) {