+++ examples/tests/plane.hpp
@@ -15,12 +15,12 @@
 
 inline void plane_test(autd3::Controller& autd) {
   autd3::SilencerConfig silencer;
-  autd.send(silencer);
+  autd.send(silencer, std::chrono::milliseconds(20));
 
   autd3::modulation::Sine m(150);  // 150Hz AM
 
   const autd3::Vector3 direction = autd3::Vector3::UnitZ();
   autd3::gain::PlaneWave g(direction);
 
-  autd.send(m, g);
+  autd.send(m, g, std::chrono::milliseconds(20));
 }